Create some of the trip routes

+++ api/controllers/TripController.js
@@ -1,4 +1,6 @@
 import Trip from '../models/TripModel.js'
+import Application from '../models/ApplicationModel.js'
+import StatusEnum from '../enum/StatusEnum.js'
 
 const listTrips = async (req, res) => {
   const filters = {}
@@ -50,6 +52,10 @@ const updateTrip = async (req, res) => {
       res.status(422).send('The trip has already been published, you can not modify it')
       return
     }
+
+    if (!newTrip.price) {
+      newTrip.price = newStages.map(stage => stage.price).reduce((totalPrice, actualPrice) => totalPrice + actualPrice, 0)
+    }
     const updatedTrip = await Trip.findOneAndUpdate({ _id: id }, newTrip, { new: true })
     res.json(updatedTrip)
   } catch (err) {
@@ -63,6 +69,16 @@ const updateTrip = async (req, res) => {
 
 const deleteTrip = async (req, res) => {
   try {
+    const trip = await Trip.findById(id)
+    if (!trip) {
+      res.status(404).send('Trip not found')
+      return
+    }
+    if (trip.publicationDate) {
+      res.status(422).send('The trip has already been published, you can not delete it')
+      return
+    }
+
     const deletionResponse = await Trip.deleteOne({ _id: req.params.id })
     if (deletionResponse.deletedCount > 0) {
       res.json({ message: 'Trip successfully deleted' })
@@ -74,4 +90,40 @@ const deleteTrip = async (req, res) => {
   }
 }
 
-export { listTrips, createTrip, readTrip, updateTrip, deleteTrip }
+const cancelTrip = async (req, res) => {
+  const { id } = req.params
+  const cancellationReason = req.body
+  try {
+    const trip = await Trip.findById(id)
+    if (!trip) {
+      res.status(404).send('Trip not found')
+      return
+    }
+    if (trip.cancellationDate) {
+      res.status(422).send('The trip has already been cancelled')
+      return
+    }
+    if (trip.publicationDate != null) {
+      res.status(422).send('The trip is not published yet')
+      return
+    }
+    const apps = await Application.find({ 'trip': trip._id, 'status': StatusEnum.ACCEPTED })
+    if (apps.length > 0) {
+      res.status(422).send('The trip has applications accepted, you can not cancel it')
+      return
+    }
+    
+    trip.cancellationDate = new Date()
+    trip.cancellationReason = cancellationReason
+    const updatedTrip = await Trip.findOneAndUpdate({ _id: id }, trip, { new: true })
+    res.json(updatedTrip)
+  } catch (err) {
+    if (err.name === 'ValidationError') {
+      res.status(422).send(err)
+    } else {
+      res.status(500).send(err)
+    }
+  }
+}
+
+export { listTrips, createTrip, readTrip, updateTrip, deleteTrip, cancelTrip }
